refactor: refactor file direction Refactor file group by functions. Especialy users and follow. Users, Follow, UserGoal entity moved.
