+++ src/main/java/goojeans/harulog/user/domain/entity/UserGoal.java
@@ -1,5 +1,7 @@
-package goojeans.harulog.domain.entity;
+package goojeans.harulog.user.domain.entity;
 
+import goojeans.harulog.domain.entity.BaseEntity;
+import goojeans.harulog.domain.entity.Category;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
