+++ src/main/java/goojeans/harulog/user/util/SocialType.java
@@ -1,0 +1,6 @@
+package goojeans.harulog.user.util;
+
+public enum SocialType {
+    KAKAO, GOOGLE, NAVER, HARU;
+
+}
